+++ src/main/java/me/drex/villagerconfig/VillagerConfig.java
@@ -5,8 +5,11 @@ import me.drex.villagerconfig.config.ConfigManager;
 import me.drex.villagerconfig.util.TradeManager;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.event.registry.DynamicRegistrySetupCallback;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.packs.PackType;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -18,11 +21,15 @@ public class VillagerConfig implements ModInitializer {
     public static final Logger LOGGER = LogManager.getLogger("VillagerConfig");
     public static final Path DATA_PATH = FabricLoader.getInstance().getConfigDir().resolve("VillagerConfig");
     public static final String MOD_ID = "villagerconfig";
-    public static final TradeManager TRADE_MANAGER = new TradeManager();
+    public static TradeManager TRADE_MANAGER;
 
     @Override
     public void onInitialize() {
-        ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(TRADE_MANAGER);
+        DynamicRegistrySetupCallback.EVENT.register(registryView -> {
+            if (TRADE_MANAGER != null) return;
+            TRADE_MANAGER = new TradeManager(registryView.asDynamicRegistryManager());
+            ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(TRADE_MANAGER);
+        });
         ConfigManager.load();
         CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> {
             VillagerConfigCommand.register(dispatcher);