+++ src/main/java/me/drex/villagerconfig/util/loot/function/EnchantRandomlyLootFunction.java
@@ -9,15 +9,14 @@ import me.drex.villagerconfig.util.loot.VCLootContextParams;
 import net.minecraft.Util;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderSet;
-import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.util.ExtraCodecs;
+import net.minecraft.core.RegistryCodecs;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.EnchantmentTags;
 import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.item.EnchantedBookItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.EnchantmentInstance;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.functions.LootItemConditionalFunction;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
@@ -26,7 +25,6 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -34,22 +32,17 @@ import java.util.Optional;
 public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
 
     private static final Logger LOGGER = LogUtils.getLogger();
-    private static final Codec<HolderSet<Enchantment>> ENCHANTMENT_SET_CODEC = BuiltInRegistries.ENCHANTMENT
-        .holderByNameCodec()
-        .listOf()
-        .xmap(HolderSet::direct, holderSet -> holderSet.stream().toList());
 
     public static final MapCodec<EnchantRandomlyLootFunction> CODEC = RecordCodecBuilder.mapCodec(
         instance -> commonFields(instance)
             .and(
                 instance.group(
-                    ENCHANTMENT_SET_CODEC.optionalFieldOf("include")
+                    RegistryCodecs.homogeneousList(Registries.ENCHANTMENT).optionalFieldOf("include")
                         .forGetter(enchantRandomlyFunction -> enchantRandomlyFunction.include),
-                    ENCHANTMENT_SET_CODEC.optionalFieldOf("exclude")
+                    RegistryCodecs.homogeneousList(Registries.ENCHANTMENT).optionalFieldOf("exclude")
                         .forGetter(enchantRandomlyFunction -> enchantRandomlyFunction.exclude),
                     Codec.INT.optionalFieldOf("min_level", 0).forGetter(enchantRandomlyLootFunction -> enchantRandomlyLootFunction.minLevel),
-                    Codec.INT.optionalFieldOf("max_level", Integer.MAX_VALUE).forGetter(enchantRandomlyLootFunction -> enchantRandomlyLootFunction.maxLevel),
-                    Codec.BOOL.fieldOf("trade_enchantments").orElse(false).forGetter(enchantRandomlyLootFunction -> enchantRandomlyLootFunction.tradeEnchantments)
+                    Codec.INT.optionalFieldOf("max_level", Integer.MAX_VALUE).forGetter(enchantRandomlyLootFunction -> enchantRandomlyLootFunction.maxLevel)
                 )
             )
             .apply(instance, EnchantRandomlyLootFunction::new)
@@ -59,15 +52,13 @@ public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
     private final Optional<HolderSet<Enchantment>> exclude;
     private final int minLevel;
     private final int maxLevel;
-    private final boolean tradeEnchantments;
 
-    EnchantRandomlyLootFunction(List<LootItemCondition> conditions, Optional<HolderSet<Enchantment>> include, Optional<HolderSet<Enchantment>> exclude, int minLevel, int maxLevel, boolean tradeEnchantments) {
+    EnchantRandomlyLootFunction(List<LootItemCondition> conditions, Optional<HolderSet<Enchantment>> include, Optional<HolderSet<Enchantment>> exclude, int minLevel, int maxLevel) {
         super(conditions);
         this.include = include;
         this.exclude = exclude;
         this.minLevel = minLevel;
         this.maxLevel = maxLevel;
-        this.tradeEnchantments = tradeEnchantments;
     }
 
     @Override
@@ -77,18 +68,9 @@ public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
             () -> {
                 boolean isBook = stack.is(Items.BOOK);
                 HolderSet<Enchantment> excluded = exclude.orElse(HolderSet.direct());
-                List<Holder.Reference<Enchantment>> list = BuiltInRegistries.ENCHANTMENT
-                    .holders()
-                    .filter(reference -> reference.value().isEnabled(context.getLevel().enabledFeatures()))
-                    .filter(reference -> reference.value().isDiscoverable())
+                List<Holder.Reference<Enchantment>> list = context.getLevel().registryAccess().registryOrThrow(Registries.ENCHANTMENT).holders()
                     .filter(reference -> isBook || reference.value().canEnchant(stack))
                     .filter(reference -> !excluded.contains(reference))
-                    .filter(reference -> {
-                        if (tradeEnchantments) {
-                            return reference.value().isTradeable();
-                        }
-                        return false;
-                    })
                     .toList();
                 return Util.getRandomSafe(list, randomSource);
             }
@@ -97,21 +79,22 @@ public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
             LOGGER.warn("Couldn't find a compatible enchantment for {}", stack);
             return stack;
         } else {
-            return enchantItem(stack, (Enchantment) ((Holder<?>) optional.get()).value(), randomSource, context);
+            return enchantItem(stack, optional.get(), randomSource, context);
         }
     }
 
-    private ItemStack enchantItem(ItemStack itemStack, Enchantment enchantment, RandomSource randomSource, LootContext context) {
+    private ItemStack enchantItem(ItemStack itemStack, Holder<Enchantment> holder, RandomSource randomSource, LootContext context) {
+        Enchantment enchantment = holder.value();
         int level = Mth.nextInt(randomSource, enchantment.getMinLevel(), enchantment.getMaxLevel());
         level = Mth.clamp(level, this.minLevel, this.maxLevel);
         if (itemStack.is(Items.BOOK)) {
             itemStack = new ItemStack(Items.ENCHANTED_BOOK);
         }
-        itemStack.enchant(enchantment, level);
+        itemStack.enchant(holder, level);
         if (context.hasParam(VCLootContextParams.NUMBER_REFERENCE)) {
             Map<String, Float> referenceProviders = context.getParamOrNull(VCLootContextParams.NUMBER_REFERENCE);
             referenceProviders.put("enchantmentLevel", (float) level);
-            referenceProviders.put("treasureMultiplier", enchantment.isTreasureOnly() ? (float) 2 : 1);
+            referenceProviders.put("treasureMultiplier", holder.is(EnchantmentTags.DOUBLE_TRADE_PRICE) ? (float) 2 : 1);
         }
         return itemStack;
     }
@@ -124,24 +107,18 @@ public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
     public static class Builder
         extends LootItemConditionalFunction.Builder<EnchantRandomlyLootFunction.Builder> {
 
-        private final List<Holder<Enchantment>> include = new ArrayList();
-        private final List<Holder<Enchantment>> exclude = new ArrayList();
+        private Optional<HolderSet<Enchantment>> include = Optional.empty();
+        private Optional<HolderSet<Enchantment>> exclude = Optional.empty();
         private int minLevel = 0;
         private int maxLevel = Integer.MAX_VALUE;
 
-        private boolean tradeEnchantments = false;
-
-        public Builder include(Enchantment... enchantments) {
-            for (Enchantment enchantment : enchantments) {
-                this.include.add(enchantment.builtInRegistryHolder());
-            }
+        public Builder include(HolderSet<Enchantment> enchantments) {
+            this.include = Optional.of(enchantments);
             return this;
         }
 
-        public Builder exclude(Enchantment... enchantments) {
-            for (Enchantment enchantment : enchantments) {
-                this.exclude.add(enchantment.builtInRegistryHolder());
-            }
+        public Builder exclude(HolderSet<Enchantment> enchantments) {
+            this.exclude = Optional.of(enchantments);
             return this;
         }
 
@@ -155,14 +132,9 @@ public class EnchantRandomlyLootFunction extends LootItemConditionalFunction {
             return this;
         }
 
-        public Builder tradeEnchantments() {
-            this.tradeEnchantments = true;
-            return this;
-        }
-
         @Override
         public @NotNull LootItemFunction build() {
-            return new EnchantRandomlyLootFunction(this.getConditions(), include.isEmpty() ? Optional.empty() : Optional.of(HolderSet.direct((include))), exclude.isEmpty() ? Optional.empty() : Optional.of(HolderSet.direct(exclude)), minLevel, maxLevel, tradeEnchantments);
+            return new EnchantRandomlyLootFunction(this.getConditions(), include, exclude, minLevel, maxLevel);
         }
 
         @Override