+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Removed `trade_enchantments` field from `enchant_randomly` loot function
+
 ### Fixed
 - Incomplete EmeraldsForVillagerTypeItem causing generation errors